+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Support the `timeout` and `proxy` options for the Symfony HTTP Client (#1084)
+
 ### 2.4.3 (2020-08-13)
 
 - Fix `Options::setEnvironment` method not accepting `null` values (#1057)