+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+### 2.5.0 (2020-09-14)
+
 - Support the `timeout` and `proxy` options for the Symfony HTTP Client (#1084)
 
 ### 2.4.3 (2020-08-13)